requirements.txt
django==4.1.2
djangorestframework==3.14
drf-yasg==1.21.4
django-filter==22.1
